+++ Sources/SwipeMenuViewController/UIViewExtension.swift
@@ -1,9 +1,0 @@
-import UIKit
-
-extension UIView {
-
-    var hasSafeAreaInsets: Bool {
-        guard #available (iOS 11, *) else { return false }
-        return safeAreaInsets != .zero
-    }
-}